feature: comicparser jsdoc
