+++ packages/comic-parser/package.json
@@ -14,13 +14,18 @@
   "bugs": "https://github.com/ridi/content-parser/issues",
   "homepage": "https://github.com/ridi/content-parser",
   "files": [
-    "lib"
+    "lib",
+    "type"
   ],
   "main": "lib/index.js",
+  "types": "type/index.d.ts",
   "dependencies": {
     "@ridi/parser-core": "0.7.0",
     "fs-extra": "^9.0.1",
     "image-size": "^0.9.1",
     "string-natural-compare": "^3.0.1"
+  },
+  "devDependencies": {
+    "@types/node": "^14.14.6"
   }
 }
