+++ packages/comic-parser/src/model/ComicReadContext.js
@@ -1,0 +1,7 @@
+import { BaseReadContext } from '@ridi/parser-core';
+
+class ComicReadContext extends BaseReadContext {
+
+}
+
+export default ComicReadContext;
