--- packages/comic-parser/src/model/ReadContext.js
@@ -1,9 +1,0 @@
-class ReadContext {
-  constructor() {
-    this.items = undefined;
-    this.entries = undefined;
-    this.options = undefined;
-  }
-}
-
-export default ReadContext;
